requirements.txt
accelerate==1.8.1
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
filelock==3.18.0
Flask==3.0.0
fsspec==2025.5.1
greenlet==3.2.3
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pgvector==0.4.1
psutil==7.0.0
psycopg2-binary==2.9.10
Pygments==2.19.2
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==13.7.0
safetensors==0.5.3
SQLAlchemy==2.0.36
sympy==1.14.0
tokenizers==0.21.2
torch==2.2.2
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
